tests/wpt/web-platform-tests/trusted-types/srcDoc.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  async_test(t => {
    var html = TrustedHTML.escape(STRINGS.unescapedHTML);

    var frame = document.createElement('iframe');
    frame.srcdoc = html;
    frame.onload = t.step_func_done(_ => {
      assert_equals(frame.srcdoc, STRINGS.escapedHTML);
    });
    document.body.appendChild(frame);
 }, "srcdoc = TrustedHTML.escape().");

  async_test(t => {
    var html = TrustedHTML.unsafelyCreate(STRINGS.unescapedHTML);

    var frame = document.createElement('iframe');
    frame.srcdoc = html;
    frame.onload = t.step_func_done(_ => {
      assert_equals(frame.srcdoc, STRINGS.unescapedHTML);
    });
    document.body.appendChild(frame);
  }, "srcdoc = TrustedHTML.unsafelyCreate().");
</script>
